+++ tests/test_decoder.py
@@ -90,15 +90,19 @@ class TestDecoder(unittest.TestCase):
                     self.assertEqual(p[1], self.d.decode_non_decimal(p[0]))
 
     def test_decode_datetime(self):
+        utc = datetime.timezone.utc
         for p in(('2001-01-01', datetime.date(2001, 1, 1)),
                  ('2001-027', datetime.date(2001, 1, 27)),
                  ('2001-027Z', datetime.date(2001, 1, 27)),
-                 ('23:45', datetime.time(23, 45)),
-                 ('01:42:57', datetime.time(1, 42, 57)),
-                 ('12:34:56.789', datetime.time(12, 34, 56, 789000)),
-                 ('2001-027T23:45', datetime.datetime(2001, 1, 27, 23, 45)),
-                 ('2001-01-01T01:34Z', datetime.datetime(2001, 1, 1, 1, 34)),
-                 ('01:42:57Z', datetime.time(1, 42, 57)),
+                 ('23:45', datetime.time(23, 45, tzinfo=utc)),
+                 ('01:42:57', datetime.time(1, 42, 57, tzinfo=utc)),
+                 ('12:34:56.789', datetime.time(12, 34, 56, 789000,
+                                                tzinfo=utc)),
+                 ('2001-027T23:45', datetime.datetime(2001, 1, 27, 23, 45,
+                                                      tzinfo=utc)),
+                 ('2001-01-01T01:34Z', datetime.datetime(2001, 1, 1, 1, 34,
+                                                         tzinfo=utc)),
+                 ('01:42:57Z', datetime.time(1, 42, 57, tzinfo=utc)),
                  ('2001-12-31T01:59:60.123Z', '2001-12-31T01:59:60.123Z'),
                  ('01:00:60', '01:00:60')):
             with self.subTest(pair=p):
@@ -149,6 +153,7 @@ class TestODLDecoder(unittest.TestCase):
 
     def test_decode_datetime(self):
         try:
+            utc = datetime.timezone.utc
             from dateutil import tz
             tz_plus_7 = tz.tzoffset('+7', datetime.timedelta(hours=7))
 
@@ -156,19 +161,21 @@ class TestODLDecoder(unittest.TestCase):
                       ('1990-158', datetime.date(1990, 6, 7)),
                       ('2001-001', datetime.date(2001, 1, 1)),
                       ('2001-01-01', datetime.date(2001, 1, 1)),
-                      ('12:00', datetime.time(12)),
-                      ('12:00:45', datetime.time(12, 0, 45)),
-                      ('12:00:45.4571', datetime.time(12, 0, 45, 457100)),
-                      ('15:24:12Z', datetime.time(15, 24, 12)),
+                      ('12:00', datetime.time(12, tzinfo=utc)),
+                      ('12:00:45', datetime.time(12, 0, 45, tzinfo=utc)),
+                      ('12:00:45.4571', datetime.time(12, 0, 45, 457100,
+                                                      tzinfo=utc)),
+                      ('15:24:12Z', datetime.time(15, 24, 12, tzinfo=utc)),
                       ('01:12:22+07',
                        datetime.time(1, 12, 22, tzinfo=tz_plus_7)),
                       ('01:12:22+7',
                        datetime.time(1, 12, 22, tzinfo=tz_plus_7)),
                       ('01:10:39.4575+07',
                        datetime.time(1, 10, 39, 457500, tzinfo=tz_plus_7)),
-                      ('1990-07-04T12:00', datetime.datetime(1990, 7, 4, 12)),
+                      ('1990-07-04T12:00', datetime.datetime(1990, 7, 4, 12,
+                                                             tzinfo=utc)),
                       ('1990-158T15:24:12Z',
-                       datetime.datetime(1990, 6, 7, 15, 24, 12)),
+                       datetime.datetime(1990, 6, 7, 15, 24, 12, tzinfo=utc)),
                       ('2001-001T01:10:39+7',
                        datetime.datetime(2001, 1, 1, 1, 10, 39,
                                          tzinfo=tz_plus_7)),